requirements.txt
# Core
pytz==2021.1
Django==3.2.5
django-configurations==2.2
gunicorn==20.1.0
newrelic==6.4.4.161
requests==2.26.0
django-redis==5.0.0
Pillow==8.3.1
drf-extra-fields==3.1.1
beautifulsoup4==4.9.3
selenium==3.141.0
geopy==2.2.0

# For the persistence stores
psycopg2-binary==2.9.1
dj-database-url==0.5.0
djongo==1.3.6
uWSGI>=2.0.19.1,<2.1

# Model Tools
django-model-utils==4.1.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Markdown==3.3.4
django-filter==2.4.0

# Developer Tools
ipdb==0.13.9
ipython==7.25.0
mkdocs==1.2.2
flake8==3.9.2

# Testing
mock==4.0.3
factory-boy==3.2.0
django-nose==1.4.7
nose-progressive==1.5.2
coverage==5.5

# Static and Media Storage
django-storages==1.11.1
boto3==1.18.2

# Documentation
drf-spectacular==0.17.3
pyyaml
uritemplate

